+++ src/lib/converter/comments/index.ts
@@ -22,6 +22,7 @@ export interface CommentParserConfig {
     inlineTags: Set<string>;
     modifierTags: Set<string>;
     jsDocCompatibility: JsDocCompatibility;
+    suppressCommentWarningsInDeclarationFiles: boolean;
 }
 
 const jsDocCommentKinds = [