+++ src/test/converter2/issues/gh2611.d.ts
@@ -1,0 +1,4 @@
+/**
+ * @tagThatIsNotDefined
+ */
+export var num: number;